OSIS-4526 move groupelementyear form to tree/attach file
