+++ program_management/tests/forms/tree/test_attach.py
@@ -24,10 +24,20 @@
 #
 ##############################################################################
 
-from django.test import SimpleTestCase
+from django.test import SimpleTestCase, TestCase
+from django.utils.translation import gettext as _
+
+from base.models.enums.education_group_types import GroupType
+from base.models.enums.link_type import LinkTypes
+from base.tests.factories.academic_year import AcademicYearFactory
+from base.tests.factories.authorized_relationship import AuthorizedRelationshipFactory
+from base.tests.factories.education_group_year import TrainingFactory, MiniTrainingFactory, EducationGroupYearFactory, \
+    GroupFactory
+from base.tests.factories.group_element_year import GroupElementYearFactory
+from base.tests.factories.learning_unit_year import LearningUnitYearFactory
 
 from program_management.ddd.domain.program_tree import ProgramTree
-from program_management.forms.tree.attach import AttachNodeForm
+from program_management.forms.tree.attach import AttachNodeForm, GroupElementYearForm
 from program_management.models.enums.node_type import NodeType
 from program_management.tests.ddd.factories.node import NodeGroupYearFactory
 
@@ -59,3 +69,130 @@ class TestAttachNodeForm(SimpleTestCase):
         form_instance = self._get_attach_node_form_instance({'link_type': 'invalid_link_type'})
         self.assertFalse(form_instance.is_valid())
         self.assertTrue(form_instance.errors['link_type'])
+
+
+class TestGroupElementYearForm(TestCase):
+    @classmethod
+    def setUpTestData(cls):
+        cls.academic_year = AcademicYearFactory()
+        cls.parent = TrainingFactory(
+            academic_year=cls.academic_year,
+            education_group_type__learning_unit_child_allowed=True
+        )
+        cls.child_leaf = LearningUnitYearFactory()
+        cls.child_branch = MiniTrainingFactory(academic_year=cls.academic_year)
+
+    def test_fields_relevant(self):
+        form = GroupElementYearForm()
+
+        expected_fields = {
+            "relative_credits",
+            "is_mandatory",
+            "block",
+            "link_type",
+            "comment",
+            "comment_english",
+            "access_condition"
+        }
+        self.assertFalse(expected_fields.symmetric_difference(set(form.fields.keys())))
+
+    def test_clean_link_type_reference_between_eg_lu(self):
+        form = GroupElementYearForm(
+            data={'link_type': LinkTypes.REFERENCE.name},
+            parent=self.parent,
+            child_leaf=self.child_leaf
+        )
+
+        self.assertTrue(form.is_valid(), form.errors)
+        self.assertTrue("link_type" not in form.fields)
+
+    def test_clean_link_type_reference_with_authorized_relationship(self):
+        AuthorizedRelationshipFactory(
+            parent_type=self.parent.education_group_type,
+            child_type=self.child_branch.education_group_type,
+        )
+        ref_group = GroupElementYearFactory(
+            parent=self.child_branch,
+            child_branch=EducationGroupYearFactory(
+                academic_year=self.academic_year,
+                education_group_type=self.child_branch.education_group_type
+            )
+        )
+        AuthorizedRelationshipFactory(
+            parent_type=self.parent.education_group_type,
+            child_type=ref_group.child_branch.education_group_type,
+        )
+
+        form = GroupElementYearForm(
+            data={'link_type': LinkTypes.REFERENCE.name},
+            parent=self.parent,
+            child_branch=self.child_branch
+        )
+
+        self.assertTrue(form.is_valid())
+
+    def test_remove_access_condition_when_not_authorized_relationship(self):
+        form = GroupElementYearForm(parent=self.parent, child_branch=self.child_branch)
+        self.assertTrue("access_condition" not in list(form.fields.keys()))
+
+    def test_only_keep_access_condition_when_parent_is_minor_major_option_list_choice(self):
+        expected_fields = ["access_condition"]
+        for name in GroupType.minor_major_option_list_choice():
+            with self.subTest(type=name):
+                parent = GroupFactory(education_group_type__name=name)
+                AuthorizedRelationshipFactory(
+                    parent_type=parent.education_group_type,
+                    child_type=self.child_branch.education_group_type
+                )
+                form = GroupElementYearForm(parent=parent, child_branch=self.child_branch)
+                self.assertCountEqual(expected_fields, list(form.fields.keys()))
+
+    def test_disable_all_fields_except_block_when_parent_is_formation_and_child_is_minor_major_option_list_choice(self):
+        expected_fields = [
+            "block"
+        ]
+        for name in GroupType.minor_major_option_list_choice():
+            with self.subTest(type=name):
+                child_branch = GroupFactory(education_group_type__name=name)
+                AuthorizedRelationshipFactory(
+                    parent_type=self.parent.education_group_type,
+                    child_type=child_branch.education_group_type
+                )
+                form = GroupElementYearForm(parent=self.parent, child_branch=child_branch)
+                enabled_fields = [name for name, field in form.fields.items() if not field.disabled]
+                self.assertCountEqual(expected_fields, enabled_fields)
+
+    def test_remove_access_condition_when_authorized_relationship(self):
+        AuthorizedRelationshipFactory(
+            parent_type=self.parent.education_group_type,
+            child_type=self.child_branch.education_group_type
+        )
+        form = GroupElementYearForm(parent=self.parent, child_branch=self.child_branch)
+        self.assertTrue("access_condition" not in list(form.fields.keys()))
+
+    def test_child_education_group_year_without_authorized_relationship_fails(self):
+        form = GroupElementYearForm(
+            data={'link_type': ""},
+            parent=self.parent,
+            child_branch=self.child_branch
+        )
+
+        self.assertFalse(form.is_valid())
+        self.assertEqual(
+            form.errors["link_type"],
+            [_("You cannot add \"%(child_types)s\" to \"%(parent)s\" (type \"%(parent_type)s\")") % {
+                 'child_types': self.child_branch.education_group_type,
+                 'parent': self.parent,
+                 'parent_type': self.parent.education_group_type,
+             }]
+        )
+
+    def test_initial_value_relative_credits(self):
+        form = GroupElementYearForm(parent=self.parent, child_branch=self.child_branch)
+        self.assertEqual(form.initial['relative_credits'], self.child_branch.credits)
+
+        form = GroupElementYearForm(parent=self.parent, child_leaf=self.child_leaf)
+        self.assertEqual(form.initial['relative_credits'], self.child_leaf.credits)
+
+        form = GroupElementYearForm(parent=self.parent)
+        self.assertEqual(form.initial['relative_credits'], None)
